desi_blog/desi_blog/urls.py
from django.conf.urls import url, include
from django.contrib import admin
from accounts.views import login_view, register_view, logout_view, activate, forgot_password, password_reset, password_change
from django.conf import settings
from django.conf.urls.static import static
from filebrowser.sites import site

urlpatterns = [
    url(r'^admin/', admin.site.urls),
    url(r'^admin/filebrowser/', include(site.urls)),
    url(r'', include('webmap.urls')),
    url(r'^login/$', login_view, name='login'),
    url(r'^register/$', register_view, name='register'),
    url(r'^forgot_password/$', forgot_password, name="forgot_password"),
    url(r'^password_change/$', password_change, name="password_change"),
    url(r'^password_reset/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$', password_reset, name="password_reset"),
    url(r'^activate/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$', activate, name='activate'),
    url(r'^logout/$', logout_view, name='logout'),
    url(r'^tinymce/', include('tinymce.urls')),
]

urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
